+++ addon/components/flexberry-objectlistview.js
@@ -6,11 +6,11 @@ import Ember from 'ember';
 import FlexberryBaseComponent from './flexberry-base-component';
 
 /**
-  Component to view list of object.
-
-  @class FlexberryObjectlistview
-  @extends FlexberryBaseComponent
-*/
+ * Component to view list of object.
+ *
+ * @class FlexberryObjectlistview
+ * @extends FlexberryBaseComponent
+ */
 export default FlexberryBaseComponent.extend({
 
   /**
@@ -38,6 +38,18 @@ export default FlexberryBaseComponent.extend({
     @default 'objectListViewRowClick'
   */
   action: 'objectListViewRowClick',
+
+  /**
+    It indicates current component mode.
+     Available values:
+     `listform` - simple list form and after row selection it has to be opened corresponding edit form;
+     `lookupform` - component is placed on lookup form and after row selection current lookup form has to be closed.
+
+    @property componentMode
+    @type String
+    @default `listform`
+  */
+  componentMode: 'listform',
 
   /**
     Default cell component that will be used to display values in columns headers.
@@ -308,13 +320,29 @@ export default FlexberryBaseComponent.extend({
   hasPreviousPage: null,
 
   /**
-    Function to determine if current page has next page.
+   * Function to determine if current page has next page.
 
     @property hasNextPage
     @type Function
     @default null
   */
   hasNextPage: null,
+
+  /**
+    Current selected record of list.
+
+    @property selectedRecord
+    @type DS.Model
+  */
+  selectedRecord: undefined,
+
+  /**
+    Set of properties to set for commponent (when it is used on lookup window).
+
+    @property customProperties
+    @type Object
+  */
+  customProperties: undefined,
 
   actions: {
     /**
@@ -349,9 +377,13 @@ export default FlexberryBaseComponent.extend({
       @param {Object} record Clicked record
     */
     objectListViewRowClick(record) {
-      let editFormRoute = this.get('editFormRoute');
-      Ember.assert('Edit form route must be defined for flexberry-objectlistview', editFormRoute);
-      this.sendAction('action', record, editFormRoute);
+      if (this.get('componentMode') === 'lookupform') {
+        this.sendAction('action', record);
+      } else {
+        let editFormRoute = this.get('editFormRoute');
+        Ember.assert('Edit form route must be defined for flexberry-objectlistview', editFormRoute);
+        this.sendAction('action', record, editFormRoute);
+      }
     },
 
     /**
@@ -464,6 +496,20 @@ export default FlexberryBaseComponent.extend({
     filterByAnyMatch(pattern) {
       throw new Error('No handler for filterByAnyMatch action set for flexberry-objectlistview. ' +
                       'Set handler like {{flexberry-objectlistview ... filterByAnyMatch=(action "filterByAnyMatch")}}.');
+    }
+  },
+
+  /**
+    An overridable method called when objects are instantiated.
+    For more information see [init](http://emberjs.com/api/classes/Ember.View.html#method_init) method of [Ember.View](http://emberjs.com/api/classes/Ember.View.html).
+   */
+  init() {
+    this._super(...arguments);
+
+    let customProperties = this.get('customProperties');
+    if (this.get('componentMode') === 'lookupform' && customProperties && typeof customProperties === 'object') {
+      // For lookup mode we allow to set properties.
+      this.setProperties(customProperties);
     }
   }
 });