+++ tests/dummy/app/controllers/ember-flexberry-dummy-suggestion-edit/new.js
@@ -1,4 +1,1 @@
-import EmberFlexberryDummySuggestionEditController from '../ember-flexberry-dummy-suggestion-edit';
-
-var EmberFlexberryDummySuggestionEditNewController = EmberFlexberryDummySuggestionEditController;
-export default EmberFlexberryDummySuggestionEditNewController;
+export { default } from '../ember-flexberry-dummy-suggestion-edit';