+++ tests/dummy/app/models/components-examples/flexberry-textbox/settings-example/base.js
@@ -1,0 +1,14 @@
+import DS from 'ember-data';
+import BaseModel from 'ember-flexberry/models/base';
+import Proj from 'ember-flexberry-data';
+
+var Model = BaseModel.extend({
+  text: DS.attr('string')
+});
+
+// Edit form projection.
+Model.defineProjection('BaseE', 'components-examples/flexberry-textbox/settings-example/base', {
+  text: Proj.attr('Text')
+});
+
+export default Model;