+++ tests/dummy/app/routes/components-examples/flexberry-textbox/settings-example.js
@@ -1,0 +1,11 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  model: function(params) {
+    var store = this.get('store');
+
+    var base = store.createRecord('components-examples/flexberry-textbox/settings-example/base', {});
+
+    return base;
+  }
+});